### 2019-03-07 #288 @harryprince
Submission：geospark
Submitting Author: Harry Zhu (@harryprince )
Repository:   [geospark](https://github.com/harryprince/geospark)
Version submitted:
Version accepted: TBD

---

```
Package: geospark
Title: Bring Local Sf to Spark
Version: 0.0.2
Author: Harry Zhu
Maintainer: Harry Zhu <7harryprince@gmail.com>
Description: This package aims at bringing local sf to distributed spark cluster for easier geocomputing.
License: Apache License
  sparklyr >= 1.0.0
```

It brings distributed GIS database to the geospatial scientist who is known PostGIS or sf R package.

1. data munging based on sparklyr package
2.  geospatial data munging

the relationship between sparklyr and geospark is platform and plugin. [GeoSpark](https://datasystemslab.github.io/GeoSpark/) is a scala package developed by [Dr. Yu](https://github.com/jiayuasu), and I found it is very convenient for the geospatial scientist who already known PostGIS or sf R package.

the geospatial scientist or transportation engineer.

The reason that I start to develop this package is making spatial operation with index easier. More generally, it is driven by bike-sharing parking business which needs to count how much bikes are within some polygons under a high presion level.

1. Better in performance：[sf](https://github.com/r-spatial/sf) and [RPostGIS](https://github.com/mablab/rpostgis)  are doing a similar job on geospatial data processing scope too, however, which is not design for big data and only running on a single machine. In that, geospark R package is a good alternative for that.
2. Broader in functionality：[sparksf](https://github.com/etiennebr/sparksf) is doing a similar thing, however, geospark package support more GIS functions than that, such as `st_isvalid` and so on.

**Better in performance**:
GeoSpark is a cluster computing system for processing large-scale spatial data with a **distributed spatial index**. Here is more detail example of performance benchmarks with PostGIS and ESRI UDF: https://github.com/harryprince/geospark#benchmarks

**Broader in functionality**:
`geospark` R package now is officially added to geospark scala package under latest [GeoSpark 1.2.0](https://github.com/DataSystemsLab/GeoSpark/tree/1.2.0-spark-2.3), which support more GIS SQL functions like `st_isvalid` and so on, you can see more features in [release notes](https://github.com/harryprince/geospark/releases).

**Great Usability and novel approach**
Here is a good example, which support `dplyr` and `sf` style grammar on spark in [geospark R package doc](https://github.com/harryprince/geospark#the-tidyverse-mode):

```
polygons_wkt <- mutate(polygons_wkt, y = st_geomfromwkt(geom, "4326"))
points_wkt <- mutate(points_wkt, x = st_geomfromwkt(geom, "4326"))

sc_res = inner_join(polygons_wkt, points_wkt, by = sql("st_contains(y, x)")) %>%
  group_by(area, state) %>%
  summarise(cnt = n())

sc_res %>%
  head()
```

```
    # Source: spark<?> [?? x 3]
    # Groups: area
      area            state   cnt
      <chr>           <chr> <dbl>
    1 texas area      TX       10
    2 dakota area     SD        1
    3 dakota area     ND       10
    4 california area CA       10
    5 new york area   NY        9
```

The reason that `geospark` R package can support dplyr and sf style grammar is  `dbplyr` can push local R functions into  `sparklyr` and parsed by the GIS SQL engine.

* https://github.com/rstudio/sparklyr/issues/1700
* https://github.com/rstudio/sparklyr/issues/1475
