### 2019-03-11 #288 @melvidoni
Thank you @harryprince for answering the questions under Scope. However, the editorial team still feels that more detail needs to be added. Consider the following questions when updating the Scope again:

- Why are you stating that target audience?
- What does sparksf compared to geospark? Why is geospark better? Adding a short example would be ideal.
- How are those issues in sparklyr related to geospark?
- Please, extend the "Description" field inside the DESCRIPTION file.
- Why does geospark imports sparklyr?

Please, tag me again when you have addressed this.
