### 2019-03-13 #288 @melvidoni
@harryprince

Your response regarding the difference between `geospark` and `sparksf` is a critique of the package author, not a comparison of the two packages. Please refrain from any such personal attacks and provide specific examples of how the packages differ based on our overlap policies (https://ropensci.github.io/dev_guide/policies.html#overlap)

On another note, with the current documentation, `geospark` appears to be a very thin R package around what are essentially Java/scala programs, but there's basically no R code to review, no R-based documentation, and no R-specific API. So far, it is the Editors' interpretation that this is essentially a port of `sf` functions into scala. This is good, but possibly not in-scope for us.

Can you please refer to this view and provide a detailed justification on why it would be in-scope? This is for Editors to make an informed decision.
