### 2019-03-14 #288 @harryprince
@melvidoni thanks for your advice, it's my first time to submit R package, I am so sorry for what I have done. I will talk with sparksf author more.
