### 2019-03-14 #288 @melvidoni
@harryprince

Do not worry, it is something that you can solve. However, please tackle the following to-do list:

- [ ] Please, remove the "peer-reviewed" badge from your repository. This should _not_ be added until your package has been reviewed and approved by rOpenSci. We are still in a pre-processing stage.

- [ ] Your answer to the question _"Are there other R packages that accomplish the same thing? If so, how does yours differ or meet our criteria for best-in-category?_" is incomplete. Please, fill it with an answer that details the following answers:

  - What means "doing it better"?
  - What tasks are improved? Why?
  - What novel or new approach did you take?

- [ ] Please, explain with detail wheather your package is just a wrap of scala functions, or has extensive R code. Do this on the _"Scope"_ section.
