### 2019-03-30 68fc908
support geospark 1.2.0
