+++ README.md
@@ -20,9 +20,9 @@ However, The Uber solution is not open source yet and Presto is not popular than
 
 In that, `geospark` R package aims at bringing local [sf](https://github.com/r-spatial/sf) functions to distributed spark mode with [GeoSpark](https://github.com/DataSystemsLab/GeoSpark) scala package.
 
-Currently, `geospark` support most of important `sf` functions in spark,
+Currently, `geospark` support the most of important `sf` functions in spark,
 here is a [summary
-comparison](https://github.com/harryprince/geospark/blob/master/Reference.md).
+comparison](https://github.com/harryprince/geospark/blob/master/Reference.md). And the `geospark` R package is keeping close with geospatial and big data community, which powered by [sparklyr](https://spark.rstudio.com), [sf](https://github.com/r-spatial/sf), [dplyr](https://db.rstudio.com/dplyr/) and [dbplyr](https://github.com/tidyverse/dbplyr).
 
 ## Installation
 
