### 2019-04-02 18d565c
add Geometry Measurement comparison
