### 2019-04-03 fa62820
Update register_gis.R
