### 2019-04-04 3d72f91
Create register_gis.Rd
