+++ man/register_gis.Rd
@@ -1,0 +1,20 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/register_gis.R
+\name{register_gis}
+\alias{register_gis}
+\title{Enable GIS SQL}
+\usage{
+register_gis(sc)
+}
+\arguments{
+\item{sc}{a sparklyr connection}
+}
+\description{
+register_gis: to enable spark geospark sql in sc, the more GIS SQL references could be found at https://github.com/harryprince/geospark/blob/master/Reference.md
+}
+\examples{
+\dontrun{
+register_gis(sc)
+point = DBI::dbGetQuery(sc,"SELECT ST_GeomFromWKT('POINT(40.7128,-74.0060)') AS geometry")
+}
+}
