### 2019-04-04 #288 @harryprince
@melvidoni I added a more detail example about how `geospark` works with `dplyr` and `sf` style grammar.
