+++ README.md
@@ -16,7 +16,7 @@ output:
 
 Goal: make tranditional GISer handle geospatial big data easier. 
 
-The origin idea comes from [Uber](https://www.oreilly.com/ideas/query-the-planet-geospatial-big-data-analytics-at-uber), which proposed a ESRI Hive UDF + Presto solution to solve large-scale geospatial data processing problem in production.
+The origin idea comes from [Uber](https://www.oreilly.com/ideas/query-the-planet-geospatial-big-data-analytics-at-uber), which proposed a ESRI Hive UDF + Presto solution to solve large-scale geospatial data processing problem with spatial index in production.
 
 However, The Uber solution is not open source yet and Presto is not popular than Spark.
 
