+++ README.Rmd
@@ -1,52 +1,92 @@
 ---
-title: "GeoSpark: R interface to GeoSpark"
+title: "GeoSpark: Bring sf to spark"
 output:
   github_document:
     fig_width: 9
     fig_height: 5
 ---
 
+![](https://image-static.segmentfault.com/101/895/1018959988-5c9809116a126)
+
 [![CRAN version](https://www.r-pkg.org/badges/version/geospark)](https://CRAN.R-project.org/package=geospark)
 [![Build Status](https://travis-ci.org/harryprince/geospark.svg?branch=master)](https://travis-ci.org/harryprince/geospark)
+[![]()](https://github.com/ropensci/software-review/issues/288)
 
 ## Introduction & Philosophy
 
-The `geospark` R package aims at bringing local [sf](https://github.com/r-spatial/sf) functions to distributed spark mode with [GeoSpark](https://github.com/DataSystemsLab/GeoSpark) scala package.
+Goal: make tranditional GISer handle geospatial big data easier. 
 
-Currently, `geospark` support most of important `sf` functions in spark, here is a [summary comparison](https://github.com/harryprince/geospark/blob/master/Reference.md).
+The origin idea comes from [Uber](https://www.oreilly.com/ideas/query-the-planet-geospatial-big-data-analytics-at-uber), which proposed a ESRI Hive UDF + Presto solution to solve large-scale geospatial data processing problem with spatial index in production.
+
+However, The Uber solution is not open source yet and Presto is not popular than Spark.
+
+In that, `geospark` R package aims at bringing local [sf](https://github.com/r-spatial/sf) functions to distributed spark mode with [GeoSpark](https://github.com/DataSystemsLab/GeoSpark) scala package.
+
+Currently, `geospark` support the most of important `sf` functions in spark,
+here is a [summary
+comparison](https://github.com/harryprince/geospark/blob/master/Reference.md). And the `geospark` R package is keeping close with geospatial and big data community, which powered by [sparklyr](https://spark.rstudio.com), [sf](https://github.com/r-spatial/sf), [dplyr](https://db.rstudio.com/dplyr/) and [dbplyr](https://github.com/tidyverse/dbplyr).
 
 ## Installation
 
-This package requires Apache Spark 2.X which you can install using `sparklyr::install_spark("2.4")`; in addition, you can install `geospark` as follows:
+This package requires Apache Spark 2.X which you can install using
+`sparklyr::install_spark("2.4")`; in addition, you can install
+`geospark` as follows:
 
-```{r eval = FALSE}
-devtools::install_github("harryprince/geospark")
+``` r
+pak::pkg_install("harryprince/geospark")
 ```
 
 ## Getting Started
 
-In this example we will join spatial data using quadrad tree indexing. First, we will initialize the `geospark` extension and connecto to Spark using `sparklyr`:
+In this example we will join spatial data using quadrad tree indexing.
+First, we will initialize the `geospark` extension and connect to Spark
+using `sparklyr`:
 
-```{r eval = FALSE}
+``` r
 library(sparklyr)
 library(geospark)
 
+conf <- spark_config()
 sc <- spark_connect(master = "local", config = conf)
+register_gis(sc)
 ```
 
-Next we will load some spatial dataset containing as polygons and points.
+Next we will load some spatial dataset containing as polygons and
+points.
 
[…]
