### 2019-04-05 #288 @melvidoni
Thanks for your submission @harryprince, and for answering all of the editors' questions. Unfortunately, `geospark`is out-of-scope for rOpenSci software-review: there is a problem of thinness, as the only R code in the package a few lines of bindings, with the rest is the scala code for the spark plugin.  The examples show how that allows interactions with other R packages.

 Thanks again for submitting `geospark` to software-review! We hope you find another venue for it.
